shit

+++ Work_2.cpp
@@ -1,13 +1,15 @@
 #include <iostream>
 #include <valarray>
+
 //y=2x/(1+sin(a+x/2))	0,7   	5,6  	0,5
 
-double myFunction(float x, float a) {
+float myFunction(float x, float a) {
     return 2 * x / (1 + sin(a + x / 2));
 }
 
 int main()
 {
+
     int i = 1;
     float x = 0.7, xStop = 5.6, xDelta = 0.5, a = 0.0;
     std::cout << "Input a: ";
